requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Babel
Flask-WTF
WTForms-Alchemy
phonenumbers
mysql-connector-python
